+++ batching-core/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.flipkart.batchingcore">
+    package="com.flipkart.batching_core">
 
     <application android:allowBackup="true" android:label="@string/app_name"
         android:supportsRtl="true">
