+++ batching-core/src/main/java/com/flipkart/batching_core/DataCollection.java
@@ -1,0 +1,52 @@
+/*
+ *  The MIT License (MIT)
+ *
+ *  Copyright (c) 2016 Flipkart Internet Pvt. Ltd.
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+
+package com.flipkart.batching_core;
+
+import java.util.Collection;
+
+/**
+ * DataCollection class
+ */
+
+public class DataCollection<T extends Data> {
+    public Collection<T> dataCollection;
+
+    public DataCollection(Collection<T> dataCollection) {
+        this.dataCollection = dataCollection;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o instanceof DataCollection) {
+            return dataCollection.equals(((DataCollection) o).dataCollection);
+        }
+        return super.equals(o);
+    }
+
+    @Override
+    public int hashCode() {
+        return dataCollection == null ? 0 : dataCollection.hashCode();
+    }
+}
