+++ batching-core/src/main/java/com/flipkart/batching_core/exception/SerializeException.java
@@ -1,0 +1,46 @@
+/*
+ *  The MIT License (MIT)
+ *
+ *  Copyright (c) 2016 Flipkart Internet Pvt. Ltd.
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+
+package com.flipkart.batching_core.exception;
+
+import java.io.IOException;
+
+/**
+ * SerializationException class that extends {@link IOException}.
+ * <p/>
+ * To get the real exception use {@link #getRealException()}.
+ */
+public class SerializeException extends IOException {
+    private final Exception realException;
+
+    public SerializeException(Exception realException) {
+        super(realException.getCause());
+        this.realException = realException;
+    }
+
+    public Exception getRealException() {
+        return realException;
+    }
+}
+
