+++ batching-core/src/test/java/com/flipkart/batching_core/Utils.java
@@ -1,0 +1,106 @@
+/*
+ *  The MIT License (MIT)
+ *
+ *  Copyright (c) 2016 Flipkart Internet Pvt. Ltd.
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+
+package com.flipkart.batching_core;
+
+import com.flipkart.batching_core.data.EventData;
+import com.flipkart.batching_core.data.Tag;
+import com.flipkart.batching_core.data.TagData;
+
+import java.util.ArrayList;
+
+/**
+ * Utils class for Test
+ */
+public class Utils {
+
+    /**
+     * Method to create fake array list of Data.
+     *
+     * @param size
+     * @return dataList
+     */
+    static Data eventData;
+
+    public static ArrayList<Data> fakeCollection(int size) {
+        ArrayList<Data> dataList = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            eventData = new EventData();
+            eventData.setEventId(System.currentTimeMillis() + System.nanoTime() + i);
+            dataList.add(eventData);
+        }
+        return dataList;
+    }
+
+    public static ArrayList<Data> fakeAdsCollection(int size) {
+        ArrayList<Data> dataList = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            eventData = new EventData();
+            eventData.setEventId(System.currentTimeMillis() + System.nanoTime() + i);
+            dataList.add(eventData);
+        }
+        return dataList;
+    }
+
+    public static ArrayList<Data> fakeDebugCollection(int size) {
+        ArrayList<Data> dataList = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            eventData = new EventData();
+            eventData.setEventId(System.currentTimeMillis() + System.nanoTime() + i);
+            dataList.add(eventData);
+        }
+        return dataList;
+    }
+
+    public static ArrayList<Data> fakeBuisnessCollection(int size) {
+        ArrayList<Data> dataList = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            eventData = new EventData();
+            eventData.setEventId(System.currentTimeMillis() + System.nanoTime() + i);
+            dataList.add(eventData);
+        }
+        return dataList;
+    }
+
+    public static ArrayList<TagData> fakeTagAdsCollection(int size) {
+        ArrayList<TagData> dataList = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            TagData eventData = new TagData(new Tag("ADS"));
+            eventData.setEventId(System.currentTimeMillis() + System.nanoTime() + i);
+            dataList.add(eventData);
+        }
+        return dataList;
+    }
+
+    public static ArrayList<TagData> fakeTagDebugCollection(int size) {
+        ArrayList<TagData> dataList = new ArrayList<>();
+        for (int i = 0; i < size; i++) {
+            TagData eventData = new TagData(new Tag("DEBUG"));
+            eventData.setEventId(System.currentTimeMillis() + System.nanoTime() + i);
+            dataList.add(eventData);
+        }
+        return dataList;
+    }
+
+}
